backend/pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["ulika <1lys.rynd@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2"
djangorestframework = "^3.16.0"
redis = "^6.0.0"
requests = "^2.32.3"
drf-yasg = "^1.21.10"
django-cors-headers = "^4.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
